added code to have watson create into db
